ML/requirements.txt
fastapi
uvicorn
pymongo[srv]
python-dotenv
pydantic
scikit-learn
xgboost
numpy
